@pivotal-cla 2021-11-19
@liuzhuni Thank you for signing the [Contributor License Agreement](https://cla.pivotal.io/sign/greenplum?repositoryId=greenplum-db/gpdb&pullRequestId=12848)!