@higuoxing 2021-11-19
This PR looks interesting. It would be great if it can be contributed to Postgres.