@gfphoenix78 2021-11-19
> we test the performance use pgbench: 1、test data gen pgbench -i -F 100 -s 11424 -p 5432-h master -U gpadmin postgres 2、test read pgbench -h master -U gpadmin -p 5432 -c 200 -j 200 -S -T 60 postgres Test platform: Kunpeng 5230 (aarch64 platform) The test result: before optimization: TPS 665 optimized: TPS 3061 This method improves performance by 360%

Do you have separate benchmark test for `CompareAndSwap()` and `compare_exchange()` in c11?