@liuzhuni 2021-11-20
we test __sync_bool_compare_and_swap() and CompareAndSwap() ,  CompareAndSwap() has better performance