@higuoxing 2021-11-20
Hi, I'm testing this patch using Postgres with my M1 Macbook.

The Postgres commit: https://github.com/postgres/postgres/commit/46d665bc26ce57b5afecbc218c8fc3c6848211d8 

The building configuration for clang:
```
./configure CC=/opt/homebrew/opt/llvm/bin/clang --target=arm64-apple-darwin --build=arm64-apple-darwin
```

The building configuration for gcc:

```
./configure CC=/opt/homebrew/bin/gcc-11 --target=arm64-apple-darwin --build=arm64-apple-darwin
```

The output of `clang -v` and `gcc -v`

```
$ /opt/homebrew/opt/llvm/bin/clang -v
Homebrew clang version 13.0.0
Target: arm64-apple-darwin20.5.0
Thread model: posix
InstalledDir: /opt/homebrew/opt/llvm/bin
```

```
$ /opt/homebrew/bin/gcc-11 -v
Using built-in specs.
COLLECT_GCC=/opt/homebrew/bin/gcc-11
COLLECT_LTO_WRAPPER=/opt/homebrew/Cellar/gcc/11.2.0_2/libexec/gcc/aarch64-apple-darwin20/11/lto-wrapper
Target: aarch64-apple-darwin20
Configured with: ../configure --prefix=/opt/homebrew/Cellar/gcc/11.2.0_2 --libdir=/opt/homebrew/Cellar/gcc/11.2.0_2/lib/gcc/11 --disable-nls --enable-checking=release --with-gcc-major-version-only --enable-languages=c,c++,objc,obj-c++,fortran --program-suffix=-11 --with-gmp=/opt/homebrew/opt/gmp --with-mpfr=/opt/homebrew/opt/mpfr --with-mpc=/opt/homebrew/opt/libmpc --with-isl=/opt/homebrew/opt/isl --with-zstd=/opt/homebrew/opt/zstd --with-pkgversion='Homebrew GCC 11.2.0_2' --with-bugurl=https://github.com/Homebrew/homebrew-core/issues --build=aarch64-apple-darwin20 --with-system-zlib --disable-multilib --with-native-system-header-dir=/usr/include --with-sysroot=/Library/Developer/CommandLineTools/SDKs/MacOSX11.sdk
Thread model: posix
Supported LTO compression algorithms: zlib zstd
gcc version 11.1.0 (Homebrew GCC 11.2.0_2)
```

Verify that clang compiles it to aarch64.
```
$ ~/.local/postgres/bin/psql  pgbenchdb -c "select version()"
                                            version
-----------------------------------------------------------------------------------------------
 PostgreSQL 15devel on aarch64-apple-darwin, compiled by Homebrew clang version 13.0.0, 64-bit
(1 row)
```

Verify that gcc compiles it to aarch64.

```
$ ~/.local/postgres/bin/psql  pgbenchdb -c "select version()"
                                                version
-------------------------------------------------------------------------------------------------------
 PostgreSQL 15devel on aarch64-apple-darwin, compiled by gcc-11 (Homebrew GCC 11.2.0_2) 11.1.0, 64-bit
(1 row)
```

Without this patch (clang).

```
$ ~/.local/postgres/bin/pgbench pgbenchdb -c 10 -j 4 -T 60
pgbench (15devel)
starting vacuum...end.
transaction type: <builtin: TPC-B (sort of)>
scaling factor: 1
query mode: simple
number of clients: 10
number of threads: 4
duration: 60 s
number of transactions actually processed: 843464
latency average = 0.711 ms
initial connection time = 6.415 ms
tps = 14057.541682 (without initial connection time)
```

Without this patch (gcc).

```
$ ~/.local/postgres/bin/pgbench pgbenchdb -c 10 -j 4 -T 60
pgbench (15devel)
starting vacuum...end.
transaction type: <builtin: TPC-B (sort of)>
scaling factor: 1
query mode: simple
number of clients: 10
number of threads: 4
duration: 60 s
number of transactions actually processed: 870703
latency average = 0.689 ms
initial connection time = 6.542 ms
tps = 14511.968933 (without initial connection time)
```

With this patch (clang).

```
$ ~/.local/postgres/bin/pgbench pgbenchdb -c 10 -j 4 -T 60
pgbench (15devel)
starting vacuum...end.
transaction type: <builtin: TPC-B (sort of)>
scaling factor: 1
query mode: simple
number of clients: 10
number of threads: 4
duration: 60 s
number of transactions actually processed: 841987
latency average = 0.713 ms
initial connection time = 6.602 ms
tps = 14033.236885 (without initial connection time)
```

With this patch (gcc).

```
$ ~/.local/postgres/bin/pgbench pgbenchdb -c 10 -j 4 -T 60
pgbench (15devel)
starting vacuum...end.
transaction type: <builtin: TPC-B (sort of)>
scaling factor: 1
query mode: simple
number of clients: 10
number of threads: 4
duration: 60 s
number of transactions actually processed: 868713
latency average = 0.691 ms
initial connection time = 7.002 ms
tps = 14478.537211 (without initial connection time)
```

It looks that this patch doesn't improve the performance? (There are some background tasks running on my laptop and I'm trying to keep the testing environment identical.

I've also tried to compile the following code snippet to assembly to show the difference.

```C
$ cat cmpswap.c
typedef int slock_t;

static __inline__ int tas1(volatile slock_t *lock) {
	return __sync_lock_test_and_set(lock, 1);
}

static __inline__ int CompareAndSwap(volatile slock_t* ptr, int old_value, int new_value)
{
	int prev;
	int temp;
	__asm__ __volatile__ (
		"0: \n\t"
		"ldaxr %w[prev], %[ptr] \n\t"
		"cmp %w[prev], %w[old_value] \n\t"
		"bne 1f \n\t"
		"stlxr %w[temp], %w[new_value], %[ptr] \n\t"
		"cbnz %w[temp], 0b \n\t"
		"1: \n\t"
		: [prev]"=&r" (prev),
		[temp]"=&r" (temp),
		[ptr]"+Q" (*ptr)
		: [old_value]"IJr" (old_value),
		[new_value]"r" (new_value)
		: "cc", "memory"
	);
	return prev;
}

static __inline__ int tas2(volatile slock_t *lock) {
	return CompareAndSwap(lock, 0, 1);
}

void foo() {
	slock_t *lock;

	tas1(lock);

	tas2(lock);
}
```

Clang's output:

```asm
$ /opt/homebrew/opt/llvm/bin/clang -S -O2 -fverbose-asm cmpswap.c -o-
	.section	__TEXT,__text,regular,pure_instructions
	.build_version macos, 11, 0
	.globl	_foo                            ; -- Begin function foo
	.p2align	2
_foo:                                   ; @foo
	.cfi_startproc
; %bb.0:
	mov	w8, #1
	swpal	w8, w9, [x8]
	; InlineAsm Start
Ltmp0:
	ldaxr	w9, [x8]
	cmp	w9, wzr
	b.ne	Ltmp1
	stlxr	w10, w8, [x8]
	cbnz	w10, Ltmp0
Ltmp1:

	; InlineAsm End
	ret
	.cfi_endproc
                                        ; -- End function
.subsections_via_symbols
```

GCC's output:

```asm
$ /opt/homebrew/bin/gcc-11 -S -O2 cmpswap.c -o-
	.arch armv8-a
	.text
	.align	2
	.p2align 4,,11
	.globl _foo
_foo:
LFB3:
	stp	x29, x30, [sp, -32]!
LCFI0:
	mov	x29, sp
	stp	x19, x20, [sp, 16]
LCFI1:
	mov	w20, 1
	mov	x19, 0
	mov	w0, w20
	mov	x1, x19
	bl	___aarch64_swp4_acq
; 11 "cmpswap.c" 1
	0:
	ldaxr w0, [x19]
	cmp w0, wzr
	bne 1f
	stlxr w1, w20, [x19]
	cbnz w1, 0b
	1:
	
; 0 "" 2
	ldp	x19, x20, [sp, 16]
	ldp	x29, x30, [sp], 32
LCFI2:
	ret

;; The eh_frame section is stripped.

LEFDE1:
	.ident	"GCC: (Homebrew GCC 11.2.0_2) 11.1.0"
	.subsections_via_symbols
```

It looks that you're implementing the TAS using a loop, hence the performance isn't as good as native atomic support from recent compilers. I also found an interesting blog post on this topic: https://blog.dbi-services.com/postgresql-on-aws-graviton2-cflags. Please correct me if I was doing something wrongly.